+++ ormar/fields/foreign_key.py
@@ -11,7 +11,7 @@ from pydantic.typing import ForwardRef, evaluate_forwardref
 from sqlalchemy import UniqueConstraint
 
 import ormar  # noqa I101
-from ormar.exceptions import RelationshipInstanceError
+from ormar.exceptions import ModelDefinitionError, RelationshipInstanceError
 from ormar.fields.base import BaseField
 
 if TYPE_CHECKING:  # pragma no cover
@@ -184,6 +184,11 @@ def ForeignKey(  # noqa CFQ002
 
     owner = kwargs.pop("owner", None)
     self_reference = kwargs.pop("self_reference", False)
+    default = kwargs.pop("default", None)
+    if default is not None:
+        raise ModelDefinitionError(
+            "Argument 'default' is not supported " "on relation fields!"
+        )
 
     if to.__class__ == ForwardRef:
         __type__ = to if not nullable else Optional[to]