+++ ormar/relations/querysetproxy.py
@@ -374,7 +374,7 @@ class QuerysetProxy:
         model = await self.queryset.get(pk=kwargs[pk_name])
         return await model.update(**kwargs)
 
-    def filter(self, **kwargs: Any) -> "QuerysetProxy":  # noqa: A003, A001
+    def filter(self, *args: Any, **kwargs: Any) -> "QuerysetProxy":  # noqa: A003, A001
         """
         Allows you to filter by any `Model` attribute/field
         as well as to fetch instances, with a filter across an FK relationship.
@@ -404,10 +404,10 @@ class QuerysetProxy:
         :return: filtered QuerysetProxy
         :rtype: QuerysetProxy
         """
-        queryset = self.queryset.filter(**kwargs)
+        queryset = self.queryset.filter(*args, **kwargs)
         return self.__class__(relation=self.relation, type_=self.type_, qryset=queryset)
 
-    def exclude(self, **kwargs: Any) -> "QuerysetProxy":  # noqa: A003, A001
+    def exclude(self, *args: Any, **kwargs: Any) -> "QuerysetProxy":  # noqa: A003, A001
         """
         Works exactly the same as filter and all modifiers (suffixes) are the same,
         but returns a *not* condition.
@@ -428,7 +428,7 @@ class QuerysetProxy:
         :return: filtered QuerysetProxy
         :rtype: QuerysetProxy
         """
-        queryset = self.queryset.exclude(**kwargs)
+        queryset = self.queryset.exclude(*args, **kwargs)
         return self.__class__(relation=self.relation, type_=self.type_, qryset=queryset)
 
     def select_related(self, related: Union[List, str]) -> "QuerysetProxy":